requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
plotly.express
scipy
altair